put support component in one dir - Other, replace all loading blocks to LoadingShema component
