--- medClinic/mainapp-ui/src/componets/LoadingSheme.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const LoadingSheme = (props) =>{
-
-    return (
-        <main className="page">
-            <section className="page__base base-block">
-                <div className="base-block__container _container">
-                    <div className="base-block__content loading-block">
-                        Loading...
-                    </div>
-                </div>
-            </section>
-        </main>
-    )
-}
-
-export default LoadingSheme
