+++ medClinic/mainapp-ui/src/componets/Other/Spoiler.jsx
